requirements.txt
yacs
tqdm
tensorboard
efficientnet_pytorch
numpy
pandas
matplotlib
